+++ src/AddEmploye.jsx
@@ -1,0 +1,12 @@
+import { useState } from "react";
+import "./App.css";
+import Form from"./Form.jsx"
+function AddEmploye() {
+  return (
+    <div id="container">
+      <Form />
+    </div>
+  );
+}
+
+export default AddEmploye;
