+++ src/MainPage.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+//import Employes from "./Employes"
+
+function MainPage() {
+  return (
+    <div>
+      <Employes />
+    </div>
+  )
+}
+
+export default MainPage
